requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
boto3==1.13.14
botocore==1.16.14
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.5
coverage==5.5
cryptography==2.9.2
decorator==5.1.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2
django-bootstrap3==15.0.0
django-bootstrap4==3.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-haystack==3.0
django-haystack-elasticsearch5==0.7
django-starfield==1.0.post2
django-storages==1.8
docutils==0.15.2
elasticsearch==5.4.0
executing==0.8.3
Faker==0.9.1
filelock==3.6.0
gunicorn==20.0.4
identify==2.4.12
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
ipython==8.3.0
jedi==0.18.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mixer==6.1.3
more-itertools==8.3.0
nodeenv==1.6.0
packaging==20.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.1
pre-commit==2.17.0
prompt-toolkit==3.0.29
psycopg2==2.8.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.12.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==6.0
requests==2.26.0
s3transfer==0.3.3
sentry-sdk==0.14.4
six==1.14.0
soupsieve==2.2.1
sqlparse==0.3.1
stack-data==0.2.0
text-unidecode==1.2
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.6
virtualenv==20.13.3
wcwidth==0.1.9
zipp==3.4.1
